refactoring: introduce TalkTypesController
- create new TalkTypesController for working with "TalkType" requests
- move createNewType() from MainPageController to TalkTypesController
- create TalkTypesControllerTest
- move tests related to TalkType requests to TalkTypesControllerTest
